# backend/app/api/static/docs/swagger.yaml
basePath: /api
definitions:
...
  repo.DuplicateOptions:
    properties:
      copyAttachments:
        type: boolean
      copyCustomFields:
        type: boolean
      copyMaintenance:
        type: boolean
      copyPrefix:
        type: string
    type: object
  repo.EntityCreate:
    properties:
      description:
        maxLength: 1000
        type: string
      entityTypeId:
        type: string
      externalId:
        maxLength: 255
        type: string
        x-nullable: true
        x-omitempty: true
      manufacturer:
        maxLength: 255
        type: string
        x-nullable: true
        x-omitempty: true
      modelNumber:
        description: |-
          Identifications — optional at create time; populated e.g. by the
          barcode product-search import flow (#1578).
        maxLength: 255
        type: string
        x-nullable: true
        x-omitempty: true
      name:
        maxLength: 255
        minLength: 1
        type: string
      parentId:
        type: string
        x-nullable: true
      quantity:
        type: number
      tagIds:
        description: Edges
        items:
          type: string
        type: array
    required:
    - name
    type: object
...
  repo.EntityListResult:
    properties:
      items:
        items:
          $ref: '#/definitions/repo.EntitySummary'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      totalPrice:
        type: number
    type: object
  repo.EntityOut:
    properties:
      archived:
        type: boolean
      assetId:
        example: "0"
        type: string
      attachments:
        items:
          $ref: '#/definitions/repo.ItemAttachment'
        type: array
      children:
        description: Container-specific fields (for entities whose entity_type.is_location
          = true)
        items:
          $ref: '#/definitions/repo.EntitySummary'
        type: array
      createdAt:
        type: string
      description:
        type: string
      entityType:
        allOf:
        - $ref: '#/definitions/repo.EntityTypeSummary'
        x-nullable: true
        x-omitempty: true
      externalId:
        type: string
      fields:
        items:
          $ref: '#/definitions/repo.EntityFieldData'
        type: array
      id:
        type: string
      imageId:
        type: string
        x-nullable: true
        x-omitempty: true
      insured:
        type: boolean
      itemCount:
        description: Container-specific (populated when querying locations)
        type: number
      lifetimeWarranty:
        description: Warranty
        type: boolean
      manufacturer:
        type: string
      modelNumber:
        type: string
      name:
        type: string
      notes:
        description: Extras
        type: string
      parent:
        allOf:
        - $ref: '#/definitions/repo.EntitySummary'
        description: Edges
        x-nullable: true
        x-omitempty: true
      purchaseDate:
        description: Purchase
        type: string
      purchaseFrom:
        type: string
      purchasePrice:
        type: number
      quantity:
        type: number
      serialNumber:
        type: string
      soldDate:
        description: Sold
        type: string
      soldNotes:
        type: string
      soldPrice:
        type: number
      soldTo:
        type: string
      syncChildEntityLocations:
        type: boolean
      tags:
        items:
          $ref: '#/definitions/repo.TagSummary'
        type: array
      thumbnailId:
        type: string
        x-nullable: true
        x-omitempty: true
      totalPrice:
        type: number
      updatedAt:
        type: string
      warrantyDetails:
        type: string
      warrantyExpires:
        type: string
    type: object
  repo.EntityPatch:
    properties:
      entityTypeId:
        type: string
        x-nullable: true
        x-omitempty: true
      id:
        type: string
      parentId:
        type: string
        x-nullable: true
        x-omitempty: true
      quantity:
        type: number
        x-nullable: true
        x-omitempty: true
      tagIds:
        items:
          type: string
        type: array
        x-nullable: true
        x-omitempty: true
    type: object
  repo.EntityPath:
    properties:
      id:
        type: string
      name:
        type: string
      type:
        $ref: '#/definitions/repo.EntityPathType'
    type: object
...
  repo.EntitySummary:
    properties:
      archived:
        type: boolean
      assetId:
        example: "0"
        type: string
      createdAt:
        type: string
      description:
        type: string
      entityType:
        allOf:
        - $ref: '#/definitions/repo.EntityTypeSummary'
        x-nullable: true
        x-omitempty: true
      id:
        type: string
      imageId:
        type: string
        x-nullable: true
        x-omitempty: true
      insured:
        type: boolean
      itemCount:
        description: Container-specific (populated when querying locations)
        type: number
      name:
        type: string
      parent:
        allOf:
        - $ref: '#/definitions/repo.EntitySummary'
        description: Edges
        x-nullable: true
        x-omitempty: true
      purchasePrice:
        type: number
      quantity:
        type: number
      soldDate:
        description: Sale details
        type: string
      tags:
        items:
          $ref: '#/definitions/repo.TagSummary'
        type: array
      thumbnailId:
        type: string
        x-nullable: true
        x-omitempty: true
      updatedAt:
        type: string
    type: object
...
  repo.EntityUpdate:
    properties:
      archived:
        type: boolean
      assetId:
        type: string
      description:
        maxLength: 1000
        type: string
      entityTypeId:
        type: string
      fields:
        items:
          $ref: '#/definitions/repo.EntityFieldData'
        type: array
      id:
        type: string
      insured:
        type: boolean
      lifetimeWarranty:
        description: Warranty
        type: boolean
      manufacturer:
        type: string
      modelNumber:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      notes:
        description: Extras
        type: string
      parentId:
        type: string
        x-nullable: true
        x-omitempty: true
      purchaseDate:
        description: Purchase
        type: string
      purchaseFrom:
        maxLength: 255
        type: string
      purchasePrice:
        type: number
        x-nullable: true
        x-omitempty: true
      quantity:
        type: number
      serialNumber:
        description: Identifications
        type: string
      soldDate:
        description: Sold
        type: string
      soldNotes:
        type: string
      soldPrice:
        type: number
        x-nullable: true
        x-omitempty: true
      soldTo:
        maxLength: 255
        type: string
      syncChildEntityLocations:
        type: boolean
      tagIds:
        description: Edges
        items:
          type: string
        type: array
      warrantyDetails:
        type: string
      warrantyExpires:
        type: string
    required:
    - name
    type: object
...
  repo.ItemAttachment:
    properties:
      createdAt:
        type: string
      id:
        type: string
      mimeType:
        type: string
      path:
        type: string
      primary:
        type: boolean
      thumbnail:
        $ref: '#/definitions/ent.Attachment'
      title:
        type: string
      type:
        type: string
      updatedAt:
        type: string
    type: object
...
  repo.TreeItem:
    properties:
      children:
        items:
          $ref: '#/definitions/repo.TreeItem'
        type: array
      id:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
...
info:
  contact:
    name: Homebox Team
    url: https://discord.homebox.software
  description: Track, Manage, and Organize your Things.
  title: Homebox API
  version: "1.0"
paths:
...
  /v1/entities:
    get:
      parameters:
      - description: search string
        in: query
        name: q
        type: string
      - description: page number
        in: query
        name: page
        type: integer
      - description: items per page
        in: query
        name: pageSize
        type: integer
      - collectionFormat: multi
        description: tags Ids
        in: query
        items:
          type: string
        name: tags
        type: array
      - collectionFormat: multi
        description: parent Ids
        in: query
        items:
          type: string
        name: parentIds
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.EntityListResult'
      security:
      - Bearer: []
      summary: Query All Entities
      tags:
      - Entities
    post:
      parameters:
      - description: Entity Data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/repo.EntityCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/repo.EntityOut'
      security:
      - Bearer: []
      summary: Create Entity
      tags:
      - Entities
  /v1/entities/{id}:
    delete:
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: Delete Entity
      tags:
      - Entities
    get:
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.EntityOut'
      security:
      - Bearer: []
      summary: Get Entity
      tags:
      - Entities
    patch:
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      - description: Entity Data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/repo.EntityPatch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.EntityOut'
      security:
      - Bearer: []
      summary: Patch Entity
      tags:
      - Entities
    put:
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      - description: Entity Data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/repo.EntityUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.EntityOut'
      security:
      - Bearer: []
      summary: Update Entity
      tags:
      - Entities
...
  /v1/entities/{id}/duplicate:
    post:
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      - description: Duplicate Options
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/repo.DuplicateOptions'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/repo.EntityOut'
      security:
      - Bearer: []
      summary: Duplicate Entity
      tags:
      - Entities
...
  /v1/entities/{id}/path:
    get:
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repo.EntityPath'
            type: array
      security:
      - Bearer: []
      summary: Get the full path of an entity
      tags:
      - Entities
  /v1/entities/export:
    get:
      responses:
        "200":
          description: text/csv
          schema:
            type: string
      security:
      - Bearer: []
      summary: Export Entities
      tags:
      - Entities
  /v1/entities/external/{id}:
    get:
      parameters:
      - description: External ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.EntityOut'
      security:
      - Bearer: []
      summary: Get Entity by External ID
      tags:
      - Entities
  /v1/entities/fields:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      security:
      - Bearer: []
      summary: Get All Custom Field Names
      tags:
      - Entities
  /v1/entities/fields/values:
    get:
      parameters:
      - description: Field name
        in: query
        name: field
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      security:
      - Bearer: []
      summary: Get All Custom Field Values
      tags:
      - Entities
  /v1/entities/import:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: CSV file to upload
        in: formData
        name: csv
        required: true
        type: file
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: Import Entities
      tags:
      - Entities
  /v1/entities/tree:
    get:
      parameters:
      - description: include items in response tree
        in: query
        name: withItems
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repo.TreeItem'
            type: array
      security:
      - Bearer: []
      summary: Get Locations Tree
      tags:
      - Entities
...
schemes:
- https
- http
securityDefinitions:
  Bearer:
    description: '"Type ''Bearer TOKEN'' to correctly set the API Key"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"